+++ tls/feature/security/src/security_default.c
@@ -157,7 +157,7 @@ int32_t SECURITY_DefaultCb(const HITLS_Ctx *ctx, const HITLS_Config *config, int
         case HITLS_SECURITY_SECOP_SIGALG_SHARED:
         case HITLS_SECURITY_SECOP_SIGALG_CHECK:
             /* Check the signature algorithm. */
-            schemeInfo = ConfigGetSignatureSchemeInfo(config, id);
+            schemeInfo = ConfigGetSignatureSchemeInfo(config, (uint16_t)id);
             if (schemeInfo != NULL && schemeInfo->secBits >= g_minBits[level - 1]) {
                 ret = SECURITY_SUCCESS;
             } else {
@@ -168,7 +168,7 @@ int32_t SECURITY_DefaultCb(const HITLS_Ctx *ctx, const HITLS_Config *config, int
         case HITLS_SECURITY_SECOP_CURVE_SHARED:
         case HITLS_SECURITY_SECOP_CURVE_CHECK:
             /* Check the group. */
-            groupInfo = ConfigGetGroupInfo(config, id);
+            groupInfo = ConfigGetGroupInfo(config, (uint16_t)id);
             if (groupInfo != NULL && groupInfo->secBits >= g_minBits[level - 1]) {
                 ret = SECURITY_SUCCESS;
             } else {
